MOVED normal slash/user/message commands into cogs
