+++ cogs/show.py
@@ -1,5 +1,5 @@
 import discord
-from discord.app import Option, OptionChoice
+from discord.app.commands import Option, OptionChoice, slash_command
 from discord.ext import commands
 from mysqldb import the_database
 from extra.useful_variables import rules
@@ -83,6 +83,19 @@ class Show(commands.Cog):
         await ctx.send("\u200b", embed=embed, view=view)
 
     # Shows the specific rule
+
+    @slash_command(name="rules", guild_ids=guild_ids)
+    @utils.is_allowed(allowed_roles)
+    async def _rules_slash(self, ctx, 
+        rule_number: Option(int, name="rule_number", description="The number of the rule you wanna show.", choices=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10], required=False), 
+        reply_message: Option(bool, name="reply_message", description="Weather the slash command should reply to your original message.", required=False, default=True)) -> None:
+        """ (MOD) Sends an embedded message containing all rules in it, or a specific rule. """
+
+        if rule_number:
+            await self._rule(ctx, rule_number, reply_message)
+        else:
+            await self._rules(ctx, reply_message)
+
     @commands.command(name="rule")
     @utils.is_allowed(allowed_roles)
     async def _rule_command(self, ctx, numb: int = None) -> None:
