+++ cogs/slothclass.py
@@ -1,10 +1,10 @@
 import discord
 from discord.ext import commands, tasks, menus
 from mysqldb import the_database
-from typing import Union, List, Any, Dict, Optional
-from extra.customerrors import MissingRequiredSlothClass, ActionSkillOnCooldown, SkillsUsedRequirement, CommandNotReady
-from extra.menu import ConfirmSkill, prompt_message, prompt_number, SlothClassPagination
-from extra.slothclasses.player import Skill, Player
+from typing import Union, List, Dict, Optional
+from extra.customerrors import ActionSkillOnCooldown, SkillsUsedRequirement, CommandNotReady
+from extra.menu import SlothClassPagination
+from extra.slothclasses.player import Skill
 from extra import utils
 import os
 
@@ -177,7 +177,7 @@ class SlothClass(*classes.values(), db_commands.SlothClassDatabaseCommands):
 
         ctx.author = member
         for c in class_commands:
-            if c.hidden or c.parent or not c.checks:
+            if c.hidden or (hasattr(c, 'parent') and c.parent) or not c.checks:
                 continue
 
             elif not [check for check in c.checks if check.__qualname__ == 'Player.skill_mark.<locals>.real_check']:
